spotify_recommender_app-master/requirements.txt
streamlit==1.29.0
pandas==1.5.3
numpy==1.24.3
scikit-learn==1.3.0
pillow==9.4.0
spotipy==2.23.0